Revert "[EC-417] Rename chrome-csv-importer to chrome-csv.importer"
This reverts commit 0f217e49d9a2c8ac3492ebedd96673425495c811.
